src/pythonDemo/requirements.txt
ontology-python-sdk